SDKTOOLS-727_mediation_out_of_the_box: android x migration
This is a long time due.
As networks like AdMob, Verizon and Vungle introduce dependencies to androidx libraries, still relying on appcompat libraries will lead to conflicts.
FairBid itself also plans on introducing a transitive dependency to google's ad identifier lib (which, in turn, depends on some androidx libs).
This migration becomes thus a necessary step.
A user should be able to compile FairBid sample app, out of box, not only to showcase FairBid stand-alone (with the unified auction) but also with all mediated networks we support.